Add submenu to plugin
