+++ config.py
@@ -1,6 +1,6 @@
 """ Module for storing plugin prefs """
 
-from calibre.utils.config import JSONConfig # type: ignore # pylint: disable=import-error no-name-in-module
+from calibre.utils.config import JSONConfig # type: ignore
 
 prefs = JSONConfig('plugins/calibre_rpc')
